updated the print_version
